use atol instead of atoi to read size_t
git-svn-id: https://svn.itwm.fhg.de/svn/SDPA@1503 c3adf61f-8968-0410-94d0-83abfdf586cf

+++ trunk/mmgr/branches/machado/fvm_pc_alloc/src/fvm/fvmConfig.c
@@ -78,7 +78,7 @@ configFile_t readConfigFile(const char *filename, const int numOptions)
 
       if(strcmp(words[0], "SHMSZ") == 0)
 	{
-	  s->shmemsize = atoi(words[1]);
+	  s->shmemsize = atol(words[1]);
 	  na++;
 	}
       else if(strcmp(words[0], "FVMSZ") == 0)
